Create blueprint for posts and shift logic to there
