+++ app/post/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+post = Blueprint('post', __name__, template_folder='post')
+
+from . import views
