+++ app/templates/base.html
@@ -36,13 +36,13 @@
                         </a>
                         <ul class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
                             <li><a class="dropdown-item" href="{{ url_for('profile.user_profile', user_id=current_user.id) }}">Мой профиль</a></li>
-                            <li><a class="dropdown-item" href="{{ url_for('profile.user_products', user_id=current_user.id) }}">Мои объявления</a></li>
+                            <li><a class="dropdown-item" href="{{ url_for('post.user_products', user_id=current_user.id) }}">Мои объявления</a></li>
                             <li><a class="dropdown-item" href="{{ url_for('profile.edit_profile', user_id=current_user.id) }}">Настройки</a></li>
                             <li><a class="dropdown-item" href="{{ url_for('auth.logout') }}">Выйти</a></li>
                         </ul>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link btn btn-primary text-light" href="{{ url_for('profile.create_product', user_id=current_user.id) }}" role="button" >Разместить объявление</a>
+                        <a class="nav-link btn btn-primary text-light" href="{{ url_for('post.create_product', user_id=current_user.id) }}" role="button" >Разместить объявление</a>
                     </li>
                     {% endif %}
                 </ul>
