+++ app/templates/profile/user.html
@@ -9,7 +9,7 @@
                 <p>{{ user.email }}</p>
                 <a href="{{ url_for('profile.edit_profile', user_id=user.id) }}" class="btn btn-light btn-sm" role="button">Редактировать профиль</a>
                 <hr>
-                <a href="{{ url_for('profile.user_products', user_id=user.id) }}">Мои объявления</a>
+                <a href="{{ url_for('post.user_products', user_id=user.id) }}">Мои объявления</a>
                 <br>
                 <a href="{{ url_for('auth.logout') }}">Выйти</a>
         </div>
